[BUG] Windows support for open file descriptor with usdcat [ENH] Dark theme support, including sticky user preference [ENH] Indentation improvements
Deprecated "-dark" flag in favor of "-theme light" or "-theme dark" to
temporarily override the sticky user theme preference.

Do not indent/unindent the ending line in a multi-line selection if the cursor
has no selection on the last line. Allow Shift-Tab to unindent the current line
when no characters are selected (previously, this shifted the widget focus).

Signed-off-by: mds-dwa <mark.sandell@dreamworks.com>