Ar2 uses the parentPath as an anchor
This updates the ArResolver2.0 code path to use the source
file as an anchor. This allows relative references to work
with non-filename asset references. Usd Manager has some
internal path manipulation to allow relative paths to work
with plain file paths, but it needs to use the resolver
plugin for non file paths. This logic already exists when
dealing with the ArResolver1.0 API, but when the 2.0 i
support was added this was overlooked.

Signed-off-by: Peter Shinners <pshinners@blizzard.com>